--- a/activerecord/lib/active_record/connection_adapters/determine_if_preparable_visitor.rb
+++ b/activerecord/lib/active_record/connection_adapters/determine_if_preparable_visitor.rb
@@ -24,6 +24,11 @@ module ActiveRecord
         @preparable = false
         super
       end
+
+      def visit_Arel_Nodes_HomogeneousIn(o, collector)
+        @preparable = false
+        super
+      end
     end
   end
 end