--- a/activerecord/lib/active_record/relation/predicate_builder/array_handler.rb
+++ b/activerecord/lib/active_record/relation/predicate_builder/array_handler.rb
@@ -21,10 +21,22 @@ module ActiveRecord
           when 0 then NullPredicate
           when 1 then predicate_builder.build(attribute, values.first)
           else
-            values.map! do |v|
-              predicate_builder.build_bind_attribute(attribute.name, v)
+            if nils.empty? && ranges.empty?
+              predicate_builder.connection.in_clause_length
+              join_name = attribute.relation.table_alias || attribute.relation.name
+              quoted_column_name = "#{predicate_builder.connection.quote_table_name(join_name)}.#{predicate_builder.connection.quote_column_name(attribute.name)}"
+              type = attribute.type_caster
+
+              casted_values = values.map do |raw_value|
+                type.serialize2(raw_value) if type.serializable?(raw_value)
+              end
+
+              casted_values.compact!
+
+              return Arel::Nodes::HomogeneousIn.new(quoted_column_name, casted_values, attribute, :in)
+            else
+              build_in(values, predicate_builder, attribute)
             end
-            values.empty? ? NullPredicate : attribute.in(values)
           end
 
         unless nils.empty?
@@ -38,6 +50,12 @@ module ActiveRecord
 
       private
         attr_reader :predicate_builder
+
+        def build_in(values, predicate_builder, attribute)
+          attribute.in values.map { |v|
+            predicate_builder.build_bind_attribute(attribute.name, v)
+          }
+        end
 
         module NullPredicate # :nodoc:
           def self.or(other)