--- a/activerecord/lib/arel/nodes.rb
+++ b/activerecord/lib/arel/nodes.rb
@@ -18,6 +18,7 @@ require "arel/nodes/false"
 # unary
 require "arel/nodes/unary"
 require "arel/nodes/grouping"
+require "arel/nodes/homogeneous_in"
 require "arel/nodes/ordering"
 require "arel/nodes/ascending"
 require "arel/nodes/descending"