@tenderlove 2020-04-21
I'm not sure that `HomogenousIn` is the best name because the collection that it wraps is not homogenous in the traditional sense.  Not all types in the collection are the same, rather they are all safe to pass to connection `quote`, and they are all terminal nodes (IOW, the `HomoegenousIn` node cannot point to a subselect, where the regular `In` node can)