@tenderlove 2020-04-21
The issue with the current `serialize` method is that it does both value validation *and* serialization (but only sometimes).  We would like to split that functionality in to two methods for this patch.

Here is an example of what I mean.  The active model integer type's `serialize` method (which can be found [here](https://github.com/rails/rails/blob/2a12b723bbc38be74275910204d29b534e80f50c/activemodel/lib/active_model/type/integer.rb#L26-L29)) looks like this:

```ruby
      def serialize(value)
        return if value.is_a?(::String) && non_numeric_string?(value)
        ensure_in_range(super)
      end
```

`ensure_in_range` will validate that the value is valid.  If it isn't, it raises an exception.  If no exception is raised, the call to `super` will actually serialize it and return the serialized form.

Another behavior that Active Record has is that we will omit values that are outside of range supported by the database.  Here is an example:

```ruby
class Post < ActiveRecord::Base
end

params = {}
params[:id] = 9999999999999999999999999999999

$-d = true
Post.where(id: [params[:id], 1234]).to_a
```

The SQL executed by the above code is:

`SELECT `posts`.* FROM `posts` WHERE `posts`.`id` IN (1234)`

Active Record strips out the very large number because the database can't handle it.

An unfortunate consequence of "raising + serializing" combined as one unit, and stripping out-of-range values is that we have to raise and catch exceptions in order to support this.  If you run the above example, you'll actually see that an `ActiveModel::RangeError` is raised in the course of running the program:

```
Exception `ActiveModel::RangeError' at /Users/aaron/git/rails/activemodel/lib/active_model/type/integer.rb:40 - 9999999999999999999999999999999 is out of range for ActiveModel::Type::Integer with limit 8 bytes
D, [2020-04-21T14:38:37.905285 #78972] DEBUG -- :   Post Load (0.5ms)  SELECT `posts`.* FROM `posts` WHERE `posts`.`id` IN (1234)
```

It would be nice if we could check that a value is valid without having to serialize it, only to get an exception.

Unfortunately the `serialize` method in Active Model is public, so we don't want to change the behavior.  Is there a different name we could use?

Second, this "raise if not valid" behavior is not consistent among all types.  For example, the `Enum` type [will not raise an exception on invalid values](https://github.com/rails/rails/blob/8544c9c23687964ab754c06a7745215a5369a4e0/activerecord/lib/active_record/enum.rb#L142-L144).

I think `serializable?` is a good method that should be in Active Model.  But I don't know what to do about `serialize` vs `serialize2`.