@tenderlove 2020-04-22
> I'm not sure that `HomogenousIn` is the best name because the collection that it wraps is not homogenous in the traditional sense. Not all types in the collection are the same, rather they are all safe to pass to connection `quote`, and they are all terminal nodes (IOW, the `HomoegenousIn` node cannot point to a subselect, where the regular `In` node can)

I was thinking about this a little more and I want to take back my statement.  We know the list is homogenous (or must be homogenous) in the "traditional sense".  All the values must have the same type.  [This line proves it](https://github.com/rails/rails/pull/39009/files#diff-3bcf02e80bc52706c3df424d2a8c4db2R28).  They all must be the same type so we can use the same type caster.

We should probably push this logic down in to the `HomogenousIn` node.